Move chrono-tz arithmetic tests to integration
